base/requirements.txt
openai
streamlit
pytest
requests
datetime
Pillow

